requirements.txt
PyYAML==5.4
Django==3.2.18
django-filter==2.4.0
djangorestframework==3.12.4
numpy==1.19.3
opencv-python==3.4.14.51
Pillow==8.3.2
pytz==2017.2
unittest-xml-reporting==3.0.2
coverage==5.1
pkg/python_ldap-3.4.0-cp39-cp39-win_amd64.whl; platform_system == 'Windows'
django_auth_ldap==2.4.0
mozilla-django-oidc==1.2.4
health-check==3.4.1
django-dramatiq==0.11.2
dramatiq[redis, watch]
unidecode
django-apscheduler